+++ include/vnx/Proxy.h
@@ -18,34 +18,12 @@
 #define INCLUDE_VNX_PROXY_H_
 
 #include <vnx/ProxyBase.hxx>
-#include <vnx/ProxyClient.hxx>
 #include <vnx/Endpoint.hxx>
-#include <vnx/TimeServer.h>
-#include <vnx/Authentication.h>
-
-#include <atomic>
 
 
 namespace vnx {
 
-/** \brief Module to connect to other processes.
- * 
- * Configuration options:
- * - \b address URL to connect to, in case given \p endpoint_ == null.
- * 		For example: "hostname" (on default port 4444), "hostname:1234" or "domain.sock".
- * - \b import_list Topics to import (ie. subscribe to) from other process.
- * - \b export_list Topics to forward to other process without them asking for it.
- * - \b forward_list Services to import (ie. forward requests to) from other process.
- * - \b tunnel_map Service tunnel map, similar to forward_list, except with a local alias (the key).
- * 		For example Hash64("other.vnx.process") => "vnx.process".
- * - \b receive_tunnel Optional tunnel pipe to forward all received samples to, instead of the local process.
- * - \b request_tunnel Optional tunnel pipe to forward all received requests to, instead of the local process.
- * - \b auto_import Flag if to automatically import all topics which are subscribed to in the local process.
- * 		(dangerous, can easily cause loops, use only if local process does not export any topics)
- * - \b time_sync Flag if to synchronize our time with the other process.
- * - \b max_queue_ms Maximum default queue length when not otherwise specified.
- * - \b max_hop_count Maximum hop count for received messages, used to defuse loops.
- */
+
 class Proxy : public ProxyBase {
 public:
 	/** \brief Create new Proxy instance.
@@ -55,115 +33,29 @@ public:
 	 * @param socket_ Already connected socket, optional
 	 */
 	Proxy(const std::string& name_, std::shared_ptr<const Endpoint> endpoint_ = 0, int socket_ = -1);
-	
+
 protected:
-	void init() override;
-	
-	void main() override;
-	
-	void login_async(const std::string &name, const std::string &password, const request_id_t& request_id) const override;
-
-	void handle(std::shared_ptr<const Sample> sample) override;
-	
-	void handle_resend(std::shared_ptr<const Sample> sample) override;
-
-	std::shared_ptr<const Return> handle(std::shared_ptr<const Request> request) override;
-	
 	bool handle(std::shared_ptr<const Return> result) override;
-	
+
 	void handle(std::shared_ptr<const FlowMessage> flow_msg) override;
 
-	void enable_import(const std::string& topic_name) override;
-	
-	void disable_import(const std::string& topic_name) override;
-	
-	void enable_export(const std::string& topic_name) override;
-	
-	void disable_export(const std::string& topic_name) override;
-	
-	void enable_forward(const std::string& service_name, const int32_t& max_queue_ms) override;
-	
-	void disable_forward(const std::string& service_name) override;
-	
-	void enable_tunnel(const Hash64& tunnel_addr, const std::string& serive_name, const int32_t& max_queue_ms) override;
-	
-	void disable_tunnel(const Hash64& tunnel_addr) override;
-	
-	void wait_on_connect_async(const request_id_t& request_id) const override;
-
-	void on_connect() override;
-	
-	void on_disconnect() override;
-	
-	void on_remote_connect(const Hash64& process_id) override;
-	
-	void on_login(const std::string &name, const std::string &password) override;
-
-	void on_remote_login(std::shared_ptr<const Session> remote_session) override;
-
 private:
-	void enable_forward(const Hash64& serive_addr, const int32_t& max_queue_ms);
-	
-	void disable_forward(const Hash64& serive_addr);
-	
-	void update_topics();
+	void send_outgoing(std::shared_ptr<const Sample> sample) override;
 
-	void print_stats();
-	
-	// to be called by read_loop() only
-	std::shared_ptr<Pipe> add_return_pipe(Hash64 src_mac, std::shared_ptr<Pipe> pipe, bool reconnect = false);
-	
-	void read_loop(std::shared_ptr<const Endpoint> endpoint);
-	
-	void close_socket();
-	
-private:
-	Hash64 service_addr;
-	Hash64 public_service_addr;
-	Hash64 remote_addr;
-	Hash64 remote_tunnel_addr;
-	Hash64 remote_process_id;
-	std::shared_ptr<const Endpoint> endpoint;
-	
-	std::unordered_map<std::string, ssize_t> import_table;			// [topic => counter]
-	std::unordered_map<std::string, ssize_t> export_table;			// [topic => counter]
-	std::unordered_map<Hash64, uint64_t> forward_table;				// [service => counter]
-	std::unordered_map<Hash64, std::shared_ptr<Pipe>> request_pipes;	// [service => pipe]
-	std::unordered_map<Hash64, Hash64> tunnel_hash_map;				// [service => service]
-	std::unordered_map<Hash128, uint64_t> channel_map;				// [(src_mac, topic) => seq_num]
-	std::unordered_set<Hash128> outgoing;							// (src_mac, dst_mac)
-	
-	std::shared_ptr<ProxyClient> remote;
-	std::shared_ptr<const TopicInfoList> topic_info;
-	Handle<TimeServer> time_server;
-	
-	bool is_connected = false;
-	bool never_connected = true;
-	std::shared_ptr<const Session> default_session;					// read-only (never modified after start)
-	std::shared_ptr<const Session> internal_session;				// read-only (never modified after start)
-	std::shared_ptr<AuthenticationServer> authentication;
-	mutable std::pair<std::string, std::string> login_credentials;
-	mutable vnx::request_t<std::shared_ptr<const Session>> login_request;
-	mutable std::vector<vnx::request_t<Hash64>> waiting_requests;
+	void send_outgoing(std::shared_ptr<const Request> request, const Hash64 &original_dst_mac) override;
 
-	SocketOutputStream stream_out;
+	void clear_outputs() override;
+
+	void read_socket_until_error(std::shared_ptr<Pipe> service_pipe) override;
+
+	void disconnect_cleanup() override;
+
+	using Super::process; // brings the other overloads into scope
+
+	void process(std::shared_ptr<FlowMessage> flow_msg) noexcept;
+
 	TypeOutput out;
-	
-	std::atomic<size_t> num_samples_send {0};
-	std::atomic<size_t> num_samples_recv {0};
-	std::atomic<size_t> num_requests_send {0};
-	std::atomic<size_t> num_requests_recv {0};
-	
-	// all below shared via mutex
-	std::mutex mutex;
-	int socket = -1;
-	std::shared_ptr<const Session> session;		// only main thread modifies
-	std::unordered_map<Hash64, std::map<uint64_t, std::shared_ptr<const Request>>> request_map;
-	
-	// all below belong to read_loop()
-	std::unordered_map<Hash64, std::shared_ptr<Pipe>> return_map;		// to keep track of return pipes
-	std::unordered_map<Hash128, std::shared_ptr<const Sample>> recv_buffer;		// last known sample per channel
-	std::unordered_set<Hash128> incoming;				// map of incoming connections (src_mac, dst_mac)
+	std::unordered_map<Hash128, uint64_t> channel_map;				// [(src_mac, topic) => seq_num]
 
 };