+++ include/vnx/vnx.h
@@ -29,6 +29,7 @@
 #include <vnx/Directory.h>
 #include <vnx/DefaultPrinter.h>
 #include <vnx/ToStringValue.h>
+#include <vnx/ToObject.h>
 #include <vnx/PrettyPrinter.h>
 #include <vnx/Process.h>
 #include <vnx/Publisher.h>